Move HpPosRecProduct class to its own files.
